separate code for license and language and add tests
