+++ src/commands/sbom/constants.ts
@@ -1,2 +1,1 @@
-export const API_ENDPOINT = 'api/v2/sbom'
-export const INTAKE_NAME = 'sbom-intake'
+export const API_ENDPOINT = 'api/v2/static-analysis-sca/dependencies'
