+++ src/commands/sbom/upload.ts
@@ -32,6 +32,7 @@ export class UploadSbomCommand extends Command {
 
   private config = {
     apiKey: process.env.DATADOG_API_KEY || process.env.DD_API_KEY,
+    appKey: process.env.DATADOG_APP_KEY || process.env.DD_APP_KEY || '',
     env: process.env.DD_ENV,
     envVarTags: process.env.DD_TAGS,
   }
@@ -69,7 +70,11 @@ export class UploadSbomCommand extends Command {
       return 1
     }
 
-    const api: (sbomPayload: ScaRequest) => AxiosPromise<AxiosResponse> = getApiHelper(this.config.apiKey)
+    // Get the API helper to send the payload
+    const api: (sbomPayload: ScaRequest) => AxiosPromise<AxiosResponse> = getApiHelper(
+      this.config.apiKey,
+      this.config.appKey
+    )
 
     const tags = await getSpanTags(this.config, this.tags)
