+++ src/constants.ts
@@ -16,6 +16,7 @@ export const DATADOG_SITES: string[] = [
 
 export const CONTENT_TYPE_HEADER = 'Content-Type'
 export const CONTENT_TYPE_VALUE_PROTOBUF = 'application/x-protobuf'
+export const CONTENT_TYPE_VALUE_JSON = 'application/json'
 
 export const METHOD_POST = 'post'
